+++ robotidy/transformers/__init__.py
@@ -39,6 +39,7 @@ TRANSFORMERS = [
     "RenameTestCases",
     "RenameKeywords",
     "ReplaceReturns",
+    "ReplaceBreakContinue",
     "InlineIf",
 ]