+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ### Transformers
 - New `ReplaceReturns` transformer that replaces return statements (such as `[Return]` setting or `Return From Keyword` keyword) with `RETURN` statement ([#231](https://github.com/MarketSquare/robotframework-tidy/issues/231))
+- New `ReplaceBreakContinue` transformer that replaces `Continue For Loop` and `Exit For Loop` keyword variants with CONTINUE and BREAK statements ([#241](https://github.com/MarketSquare/robotframework-tidy/issues/241))
 - New `InlineIf` transformer that replaces IF blocks with inline IF ([#230](https://github.com/MarketSquare/robotframework-tidy/issues/230))
 - Add `min_width` parameter to `AlignSettingsSection` transformer. It allows to set minimal width of data column ([#242](https://github.com/MarketSquare/robotframework-tidy/issues/242))
 - Add `min_width` parameter to `AlignVariablesSection` transformer. It allows to set minimal width of data column ([#242](https://github.com/MarketSquare/robotframework-tidy/issues/242))